Modify how systemd detection works to make it more lenient.
The current systemd detection routines assume that the operating
environment in which services are installed will be the same as the
one in which the service will run in.  This is not true in all cases.

For example, the ubuntu 18.04 install environment uses busybox init
as the init system, and the final install environment uses systemd.